[#179]:svarga:refactor, add HDF5 1.12 reference compatibility layer
