+++ h5cpp/H5Dread.hpp
@@ -6,6 +6,7 @@
 #pragma once
 #include "H5capi.hpp"
 #include "H5Dopen.hpp" // be sure this precedes error handling macro-s !!!
+#include "H5Rreference.hpp"
 #include <string>
 #include <vector>
 #include <stdexcept>
@@ -283,11 +284,7 @@ namespace h5 {
 			for(size_t i=0; i<nelem; i++)
 					if( ptr[i] != nullptr )
 						ref[i] = std::string( ptr[i] );
-#if H5_VERSION_GE(1,12,0)
-		H5Treclaim(mem_type, mem_space, H5P_DEFAULT, ptr);
-#else
-		H5Dvlen_reclaim(mem_type, mem_space, H5P_DEFAULT, ptr);
-#endif
+		h5::impl::reference::reclaim(mem_type, mem_space, H5P_DEFAULT, ptr);
 		free(ptr);
 		return ref;
 	}
