+++ test/CMakeLists.txt
@@ -55,6 +55,7 @@ add_test_case(H5Adelete)
 add_test_case(H5compat)
 add_test_case(H5Dgather)
 add_test_case(H5meta)
+add_test_case(H5Rreference)
 add_test_case(H5Tmeta)
 add_test_case(H5Dio)
 add_test_case(H5Eall)
